Add attach and detach to class RecruiterController extends Controller

+++ database/migrations/2023_03_30_091116_create_matches_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->unsignedBigInteger('recruiter_id');
             $table->foreign('recruiter_id')->references('id')->on('recruiters');
 
-            $table->float('afinity',3,2);    
+            $table->decimal('afinity',5,2);    
 
             $table->timestamps();
         });
